Medicine last day + medicine details page edit
Since period of the medicine intake depends on the interval the user inputs, the last day is not inputted by the user, but rather calculated from the interval/type/number of repeated notifications from the first day. The calendar fragment will display the medicine on the days of the intake duration (firstday-lastday) accordingly, taking the date of the calculated lastday rather than an inputted last day.
Also, the medicine reminder details page has been edited to include the timely interval between each reminder of that medicine. For example, if the user wants the reminder to repeat every 6 hours, "every 6 hour(s)" will be displayed next to the first reminder time in the medicine details page.